(ts) @curi/redux
This commit also removes the config reducer because you shouldn't be a reason that you create a new configuration object while your app is running.
