--- packages/redux/.babelrc
@@ -1,3 +1,0 @@
-{
-  "presets": [ "./babel/config" ]
-}
